+++ src/index.js
@@ -1,123 +1,130 @@
 const TESTING = global.COMB !== undefined;
-
-if (!TESTING)
+if (!TESTING) {
   COMB = require('@holo-host/comb').COMB;
+}
 
 const { EventEmitter } = require('events');
 
-function makeUrlAbsolute (url) {
+function makeUrlAbsolute(url) {
   return new URL(url, window.location).href
 }
 
-class Connection extends EventEmitter {
-
-  constructor(url, signalCb, opts) {
+/**
+ * The `WebSdkApi` class is a wrapper around the COMB.js library that provides a JavaScript API for Holo-Hosted web apps to call Holochain. 
+ * @param child - The child process connecting to Chaperone that is being monitored.
+ */
+class WebSdkApi extends EventEmitter {
+  constructor(child) {
     super();
-
-    const hostname = window.location.hostname;
-    this.chaperone_url = new URL(url || `http://${hostname}:24273`);
-    if (opts !== undefined) {
-      if (opts.logo_url !== undefined) {
-        this.chaperone_url.searchParams.set("logo_url", makeUrlAbsolute(opts.logo_url))
-      }
-      if (opts.app_name !== undefined) {
-        this.chaperone_url.searchParams.set("app_name", opts.app_name)
-      }
-      if (opts.info_link !== undefined) {
-        this.chaperone_url.searchParams.set("info_link", opts.info_link)
-      }
-      if (opts.publisher_name !== undefined) {
-        this.chaperone_url.searchParams.set("publisher_name", opts.publisher_name)
-      }
-      if (opts.registration_server !== undefined) {
-        this.chaperone_url.searchParams.set("registration_server_url", makeUrlAbsolute(opts.registration_server.url))
-        this.chaperone_url.searchParams.set("registration_server_payload", JSON.stringify(opts.registration_server.payload))
-      }
-      if (opts.skip_registration !== undefined) {
-        this.chaperone_url.searchParams.set("skip_registration", opts.skip_registration)
-      }
-    }
-
-    this.waiting = [];
-    this.child = null;
-    this.signalCb = signalCb
-    this.connecting = this.connect();
-  }
-
-  ready() {
-    return new Promise((resolve, reject) => {
-      this.connecting.catch(reject)
-      this.child !== null
-        ? resolve()
-        : this.waiting.push(resolve);
+    this.available = null;
+    this.child = child;
+    child.msg_bus.on("alert", (event, ...args) => this.emit(event));
+    child.msg_bus.on("signal", signal => this.emit('signal', signal));
+    child.msg_bus.on("unrecoverable-agent-state", () => this.emit('unrecoverable-agent-state'));
+    child.msg_bus.on("unavailable", () => this.emit('unavailable'));
+    child.msg_bus.on("available", () => {
+      this.available()
+      this.emit('available')
     });
   }
 
-  async connect() {
+  /* The `ready` function is a promise that is resolved when the WebSDK is ready to be used. */
+  ready = () => {
+    return new Promise((resolve, reject) => {
+      this.available = resolve
+    });
+  }
+
+  /**
+   * The `static connect` function is a helper function that connects to the Chaperone server and returns a WebSdkApi object
+   * @param [] 
+   * - chaperoneUrl: The URL of the Chaperone server.
+   * - authFormCustomization: The optional app customizations to display when authorizing the user 
+   * @returns The `connect` function returns a `WebSdkApi` object.
+   */
+  static connect = async ({ chaperoneUrl, authFormCustomization: authOpts } = {}) => {
+    const hostname = window.location.hostname;
+    const url = new URL(chaperoneUrl || `http://${hostname}:24273`);
+    if (authOpts !== undefined) {
+      if (authOpts.logoUrl !== undefined) {
+        url.searchParams.set("logo_url", makeUrlAbsolute(authOpts.logoUrl))
+      }
+      if (authOpts.appName !== undefined) {
+        url.searchParams.set("app_name", authOpts.appName)
+      }
+      if (authOpts.infoLink !== undefined) {
+        url.searchParams.set("info_link", authOpts.infoLink)
+      }
+      if (authOpts.publisherName !== undefined) {
+        url.searchParams.set("publisher_name", authOpts.publisherName)
+      }
+      if (authOpts.anonymousAllowed !== undefined) {
+        url.searchParams.set("anonymous_allowed", authOpts.anonymousAllowed)
+      }
+      if (authOpts.registrationServer !== undefined) {
+        url.searchParams.set("registration_server_url", makeUrlAbsolute(authOpts.registrationServer.url))
+        url.searchParams.set("registration_server_payload", JSON.stringify(authOpts.registrationServer.payload))
+      }
+      if (authOpts.skipRegistration !== undefined) {
+        url.searchParams.set("skip_registration", authOpts.skipRegistration)
+      }
+    }
+
+    let child
     try {
-      this.child = await COMB.connect(this.chaperone_url.href, 60000, this.signalCb);
+      child = await COMB.connect(url.href, 60000);
     } catch (err) {
       if (err.name === "TimeoutError")
         console.log("Chaperone did not load properly. Is it running?");
       throw err;
     }
 
-    let f;
-    while (f = this.waiting.shift()) {
-      f();
+    const webSdkApi = new WebSdkApi(child);
+
+    // Set styles and history props when in production mode
+    // Note: Set styles and history props only when in production mode
+    if (!TESTING) {
+      webSdkApi.iframe = document.getElementsByClassName("comb-frame-0")[0];
+      webSdkApi.iframe.setAttribute('allowtransparency', 'true');
+      const style = webSdkApi.iframe.style;
+      style.zIndex = "99999999";
+      style.width = "100%";
+      style.height = "100%";
+      style.position = "absolute";
+      style.top = "0";
+      style.left = "0";
+      style.display = "none";
+      window.addEventListener('popstate', (event) => {
+        if (event.state === "_web_sdk_shown") {
+          history.back()
+        } else {
+          webSdkApi.iframe.style.display = "none"
+        }
+      });
     }
 
-    if (TESTING)
-      return;
+    // Note: Based on discussed model, chaperone either returns null or an error message (string)
+    const error_message = await child.call("handshake")
+    if (error_message) {
+      webSdkApi.iframe.display = "none"
+      throw new Error(error_message)
+    }
 
-    // Alerts:
-    //   signin		- emitted when the user completes a successful sign-in
-    //   signup		- emitted when the user completes a successful sign-up
-    //   signout		- emitted when the user competes a successful sign-out
-    //   canceled		- emitted when the user purposefully exits sign-in/up
-    //   connected		- emitted when the connection is opened
-    //   disconnected	- emitted when the connection is closed
-    this.child.msg_bus.on("alert", (event, ...args) => {
-      this.emit(event);
-    });
-
-    this.iframe = document.getElementsByClassName("comb-frame-0")[0];
-    this.iframe.setAttribute('allowtransparency', 'true');
-
-    const style = this.iframe.style;
-    style.zIndex = "99999999";
-    style.width = "100%";
-    style.height = "100%";
-    style.position = "absolute";
-    style.top = "0";
-    style.left = "0";
-    style.display = "none";
-
-    window.addEventListener('popstate', (event) => {
-      if (event.state === "_web_sdk_shown") {
-        history.back()
-      } else {
-        this.iframe.style.display = "none"
-      }
-    })
+    return webSdkApi
   }
 
-  async agentInfo() {
-    const response = await this.child.call("agentInfo");
-    return response;
-  }
+  zomeCall = async (...args) => await this.child.call("zomeCall", ...args)
 
-  async zomeCall(...args) {
-    const response = await this.child.call("zomeCall", ...args);
-    return response;
-  }
+  appInfo = async (...args) => await this.child.call("appInfo", ...args)
 
-  async appInfo(...args) {
-    const response = await this.child.call("appInfo", ...args);
-    return response;
-  }
+  cellData = async (args) => await this.child.call("cellData", ...args)
 
-  async signUp(opts) {
+  stateDump = async () => await this.child.call("stateDump")
+
+  /* The `signUp` function is called by the `signUp` button in the UI. The `signUp`
+  function is async, so it returns a promise. The promise is resolved when the user is signed up. The
+  promise is rejected if a Holo error occurs or the user cancels the sign up process (when the cancellable opt is provided). */
+  signUp = async (opts) => {
     const { cancellable = true } = opts || {}
     if (cancellable) {
       history.pushState("_web_sdk_shown", "")
@@ -134,7 +141,10 @@ class Connection extends EventEmitter {
     return result;
   }
 
-  async signIn(opts) {
+  /* The `signIn` function is called by the `signIn` button in the UI. The `signIn`
+  function is async, so it returns a promise. The promise is resolved when the user is signed in. The
+  promise is rejected if a Holo error occurs or the user cancels the sign in process (when the cancellable opt is provided). */
+  signIn = async (opts) => {
     const { cancellable = true } = opts || {}
     if (cancellable) {
       history.pushState("_web_sdk_shown", "")
@@ -151,19 +161,7 @@ class Connection extends EventEmitter {
     return result;
   }
 
-  async signOut() {
-    return await this.child.run("signOut");
-  }
-
-  async holoInfo() {
-    return await this.child.run("holoInfo");
-  }
+  signOut = async () => await this.child.run("signOut")
 }
 
-Connection.AUTONOMOUS = 1;
-Connection.HOSTED_ANONYMOUS = 2;
-Connection.HOSTED_AGENT = 3;
-
-module.exports = {
-  Connection,
-};
+module.exports = WebSdkApi