+++ tests/mock_comb.js
@@ -1,9 +1,18 @@
+const { inspect } = require("util");
 
 let next_response;
+let event_listeners = {}
 
 global.COMB = {
   async connect() {
     return Promise.resolve({
+      // Test event listener set with callback
+      msg_bus: {
+        on(event, callback) {
+          event_listeners[event] = callback
+          return
+        }
+      },
       call() {
         if (next_response === undefined)
           return `{"Err":"Next response is undefined"}`
@@ -12,7 +21,7 @@ global.COMB = {
           next_response = undefined;
           return value;
         }
-      }
+      },
     });
   }
 }
@@ -20,5 +29,15 @@ global.COMB = {
 module.exports = {
   nextResponse(value) {
     next_response = value;
+  },
+  triggerEvent(event_name) {
+    const event_callback = event_listeners[event_name]
+    if (!event_callback) {
+      throw new Error(`Expected to find event listener ${event_name}, but none found.  Current event listeners: ${inspect(event_listeners)}`)
+    }
+    setTimeout(() => {
+      event_callback()
+    }, 0)
+    return
   }
 }